API Testing, Exception Handling

+++ src/main/java/com/project/application/BankingApp/Service/ServiceImplementation/AccountImplementation.java
@@ -1,10 +1,13 @@
 package com.project.application.BankingApp.Service.ServiceImplementation;
 
 import com.project.application.BankingApp.Entity.Account;
+import com.project.application.BankingApp.Exception.AccountNotFound;
+import com.project.application.BankingApp.Exception.InsufficientBalance;
 import com.project.application.BankingApp.Repository.AccountRepository;
 import com.project.application.BankingApp.Service.AccountService;
 import com.project.application.BankingApp.dto.AccountDto;
 import org.modelmapper.ModelMapper;
+import org.modelmapper.internal.bytebuddy.implementation.bytecode.Throw;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -30,12 +33,20 @@ public class AccountImplementation implements AccountService {
 
     @Override
     public AccountDto GetAccount(Long id) {
+        if(!accountRepository.findById(id).isPresent()){
+            throw new AccountNotFound(id);
+        }
         return this.modelMapper.map(this.accountRepository.findById(id), AccountDto.class);
     }
 
     @Override
     public String DeleteAccount(Long id) {
-        accountRepository.deleteById(id);
+        if(!accountRepository.findById(id).isPresent()){
+            throw new AccountNotFound(id);
+        }
+        else{
+            accountRepository.deleteById(id);
+        }
         return "Account with id" + id + "is deleted";
     }
 
@@ -47,6 +58,28 @@ public class AccountImplementation implements AccountService {
             AccountDtoList.add(modelMapper.map(a, AccountDto.class));
         }
         return AccountDtoList;
+    }
+
+    @Override
+    public AccountDto DepositAmount(double DepositAmount, Long id) {
+        Account account = accountRepository.findById(id).orElseThrow(() -> new RuntimeException("Account does not exist"));
+        double BalanceAfterDeposit = DepositAmount + account.getAccountBalance();
+        account.setAccountBalance(BalanceAfterDeposit);
+        accountRepository.save(account);
+        return this.modelMapper.map(account, AccountDto.class);
+    }
+
+    @Override
+    public AccountDto WithdrawAmount(double WithdrawAmount, Long id) {
+        Account account = accountRepository.findById(id).orElseThrow(() -> new RuntimeException("Account does not exist"));
+        if(account.getAccountBalance() <= 100 || WithdrawAmount>=account.getAccountBalance()){
+            throw new InsufficientBalance(account.getAccountBalance());
+        } else{
+            double AmountLeft = account.getAccountBalance() - WithdrawAmount;
+            account.setAccountBalance(AmountLeft);
+            accountRepository.save(account);
+        }
+        return this.modelMapper.map(account, AccountDto.class);
     }
